adding delete saved job functionality

+++ src/app/features/modal-job-details/modal-job-details.component.ts
@@ -3,6 +3,8 @@ import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { DialogModule } from 'primeng/dialog';
 import { Job } from '../../core/models/jobs';
 import { NgbModal, NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { ToastModule } from 'primeng/toast';
+import { MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-modal-job-details',
@@ -10,19 +12,23 @@ import { NgbModal, NgbModule } from '@ng-bootstrap/ng-bootstrap';
   imports: [
     CommonModule,
     DialogModule,
-    NgbModule
+    NgbModule,
+    ToastModule
   ],
   templateUrl: './modal-job-details.component.html',
-  styleUrl: './modal-job-details.component.scss'
+  styleUrls: ['./modal-job-details.component.scss'],
+  providers: [MessageService]  // Add MessageService here
 })
 export class ModalJobDetailsComponent implements OnInit {
   @ViewChild('content', { static: false }) content!: ElementRef;
-  constructor(private modalService: NgbModal) { }
+  @Input() job!: Job;
+  display: boolean = false;
+
+  constructor(private modalService: NgbModal, private messageService: MessageService) {}
+
   ngOnInit(): void {
     console.log(this.job);
   }
-  @Input() job: any;
-  display: boolean = false;
 
   showDialog() {
     this.open();
@@ -34,19 +40,23 @@ export class ModalJobDetailsComponent implements OnInit {
 
   saveJob(newJob: Job) {
     let savedJobs = localStorage.getItem('savedJobs');
-  
+
     // If savedJobs doesn't exist, create an empty array
     if (!savedJobs) {
       savedJobs = '[]';
     }
-  
-    // Parse the savedJobs and add the new job
-    const savedJobsArray: Job[] = JSON.parse(savedJobs);
-    savedJobsArray.push(newJob);
-  
-    // Save the updated array back to localStorage
-    localStorage.setItem('savedJobs', JSON.stringify(savedJobsArray));
-    this.modalService.dismissAll()
-  }  
 
+    // Parse the savedJobs and check if the job already exists
+    const savedJobsArray: Job[] = JSON.parse(savedJobs);
+    const jobExists = savedJobsArray.some(job => job.id === newJob.id);
+
+    if (jobExists) {
+      this.messageService.add({ severity: 'warn', summary: 'Warning', detail: 'Job already saved' });
+    } else {
+      savedJobsArray.push(newJob);
+      localStorage.setItem('savedJobs', JSON.stringify(savedJobsArray));
+      this.messageService.add({ severity: 'success', summary: 'Success', detail: 'Job saved successfully' });
+    }
+    this.modalService.dismissAll();
+  }
 }
